Integrate arguments building to Action
