+++ lib/problems/components/context.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Problems
+  # Action execution context.
+  class Context
+  end
+end
