+++ test/common/validator_test.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'test_helper'
 require 'problems/common/validator'
 
 class ValidatorTest < Minitest::Test
